app/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# TTS - Piper
piper-tts==1.2.0

# STT - Faster Whisper (minimal)
faster-whisper==1.1.0
